fix: harden session reset detection
